# server/go2rtc/api/openapi.yaml
openapi: 3.1.0
info:
  title: go2rtc
  license: { name: MIT,url: https://opensource.org/licenses/MIT }
  version: 1.0.0
  contact: { url: https://github.com/AlexxIT/go2rtc }
  description: |
    Ultimate camera streaming application with support RTSP, RTMP, HTTP-FLV, WebRTC, MSE, HLS, MP4, MJPEG, HomeKit, FFmpeg, etc.
servers:
  - url: http://localhost:1984
components:
  parameters:
    stream_src_path:
      name: src
      in: path
      description: Source stream name
      required: true
      schema: { type: string }
      example: camera1
    stream_dst_path:
      name: dst
      in: path
      description: Destination stream name
      required: true
      schema: { type: string }
      example: camera1
...
paths:
...
  /api/stream.mjpeg?src={src}:
    get:
      summary: Get stream in MJPEG format
      description: "[Module: MJPEG](https://github.com/AlexxIT/go2rtc#module-mjpeg)"
      tags: [ Consume stream ]
      parameters:
        - $ref: "#/components/parameters/stream_src_path"
      responses:
        200:
          description: ""
          content: { multipart/x-mixed-replace: { example: "" } }
  /api/frame.jpeg?src={src}:
    get:
      summary: Get snapshot in JPEG format
      description: "[Module: MJPEG](https://github.com/AlexxIT/go2rtc#module-mjpeg)"
      tags: [ Snapshot ]
      parameters:
        - $ref: "#/components/parameters/stream_src_path"
      responses:
        200:
          description: ""
          content: { image/jpeg: { example: "" } }
...
  /api/stream.mjpeg?dst={dst}:
    post:
      summary: Post stream in MJPEG format
      description: "[Incoming sources](https://github.com/AlexxIT/go2rtc#incoming-sources)"
      tags: [ Produce stream ]
      parameters:
        - $ref: "#/components/parameters/stream_dst_path"
      responses:
            default:
              description: Default response
...